migrate and fix storybook
