+++ src/lib/components/Tip/Tip.stories.jsx
